# openapi.yaml
openapi: 3.1.0
info:
    title: VoiceBlender API
    description: |
        VoiceBlender bridges SIP and WebRTC voice calls with multi-party audio mixing, real-time speech-to-text, text-to-speech, AI agent integration, recording, and webhook-based event delivery.
    x-config-vars:
        - name: INSTANCE_ID
          default: (auto-generated UUID)
          description: Instance identifier included in all API responses and webhook events
        - name: HTTP_ADDR
          default: :8080
          description: REST API listen address
        - name: SIP_BIND_IP
          default: 127.0.0.1
          description: IPv4 address advertised in SDP, Contact, and Via headers (and used as the listen address when SIP_LISTEN_IP is empty)
        - name: SIP_LISTEN_IP
          default: (same as SIP_BIND_IP)
          description: UDP socket bind IP. Accepts 127.0.0.1, 0.0.0.0, ::, or any literal v4/v6 address
        - name: SIP_BIND_IPV6
          default: (empty = v4-only)
          description: IPv6 address advertised in SDP/Contact/Via for IPv6 calls. Set this for IPv6-only or dual-stack deployments
        - name: SIP_LISTEN_IPV6
          default: (same as SIP_BIND_IPV6)
          description: Optional separate IPv6 socket bind address (used when running with both 0.0.0.0 and a specific v6 literal)
        - name: SIP_PORT
          default: "5060"
          description: SIP listen port
        - name: SIP_HOST
          default: voiceblender
...
        - name: ICE_SERVERS
...
servers:
...
paths:
...
    /legs/{id}/rtt:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: sendRTT
            summary: Send Real-Time Text (T.140) on a SIP leg
            description: Sends UTF-8 text on the leg's RTT (T.140 / RFC 4103) media stream. Requires that the SDP offer/answer agreed on an m=text section with the remote UA. Enable RTT on the server with RTT_ENABLED=true.
            tags:
                - Legs
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RTTRequest'
            responses:
                '200':
                    description: Text sent
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '400':
                    description: Invalid JSON or empty text
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '409':
                    description: RTT was not negotiated for this leg
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                '500':
                    description: Send failed
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/rtt/accept:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: acceptRTTLeg
            summary: Enable RTT reception on a leg
            description: Allow this leg to receive RTT text broadcast from other legs in the same room and to emit rtt.received events for incoming text. Default state for new legs.
            tags:
                - Legs
            responses:
                '200':
                    description: RTT reception enabled
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
    /legs/{id}/rtt/reject:
        parameters:
            - $ref: '#/components/parameters/LegId'
        post:
            operationId: rejectRTTLeg
            summary: Disable RTT reception on a leg
            description: Block this leg from receiving RTT text broadcast from other legs in the same room and suppress rtt.received events for this leg.
            tags:
                - Legs
            responses:
                '200':
                    description: RTT reception disabled
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/StatusResponse'
                '404':
                    description: Leg not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
...
components:
    parameters:
        LegId:
            name: id
            in: path
            required: true
            schema:
                type: string
            description: Leg ID
...
    schemas:
...
        Error:
            type: object
            properties:
                instance_id:
                    type: string
                    description: Instance identifier
                error:
                    type: string
                    description: Error message
            required:
                - error
        StatusResponse:
            type: object
            properties:
                instance_id:
                    type: string
                    description: Instance identifier
                status:
                    type: string
            required:
                - status
...
        RTTRequest:
            type: object
            properties:
                text:
                    type: string
                    description: UTF-8 text to send. May be one or more characters and may include T.140 control codes (e.g. backspace U+0008, CR/LF).
            required:
                - text
...